requirements.txt
opencv-python-headless
dlib
numpy
ultralytics
mistralai
Pillow
requests
SpeechRecognition
face-recognition
picamera2
moondream
python-dotenv